% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_legends.R
\name{create_circle_legend}
\alias{create_circle_legend}
\title{Create a circle legend}
\usage{
create_circle_legend(title, palette, position_args = position_arguments())
}
\description{
Create a circle legend
}
\examples{
title <- "Greys"
palette <- funkyheatmap:::default_palettes$numerical$Greys
create_circle_legend(title, palette)
}
