% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_legends.R
\name{create_funkyrect_legend}
\alias{create_funkyrect_legend}
\title{Create a funkyrect legend}
\usage{
create_funkyrect_legend(title, palette, position_args = position_arguments())
}
\description{
Create a funkyrect legend
}
\examples{
title <- "Greys"
palette <- funkyheatmap:::default_palettes$numerical$Greys
create_funkyrect_legend(title, palette)
}
