% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_legends.R
\name{create_generic_geom_legend}
\alias{create_generic_geom_legend}
\title{Create a generic geom legend (for circles, rects, and funkyrects)}
\usage{
create_generic_geom_legend(
  title,
  palette,
  geom = c("circle", "rect", "funkyrect"),
  position_args = position_arguments()
)
}
\arguments{
\item{title}{The name of the palette}

\item{palette}{The palette}

\item{geom}{Which geom to draw.}

\item{position_args}{Sets parameters that affect positioning within a
plot, such as row and column dimensions, annotation details, and the
expansion directions of the plot. See \code{position_arguments()} for more information.}
}
\description{
Create a generic geom legend (for circles, rects, and funkyrects)
}
\examples{
title <- "Greys"
palette <- funkyheatmap:::default_palettes$numerical$Greys
geom <- "circle"
create_scaling_geom_legend(title, palette, geom)
}
