% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_legends.R
\name{create_pie_legend}
\alias{create_pie_legend}
\title{Create a pie legend}
\usage{
create_pie_legend(title, palette, position_args = position_arguments())
}
\arguments{
\item{title}{The name of the palette}

\item{palette}{The palette}

\item{position_args}{Sets parameters that affect positioning within a
plot, such as row and column dimensions, annotation details, and the
expansion directions of the plot. See \code{position_arguments()} for more information.}
}
\description{
Create a pie legend
}
\examples{
title <- "Greys"
palette <- c(One = "#FF0000", Two = "#00FF00", Three = "#0000FF")
create_pie_legend(title, palette)
}
