% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynbenchmark_data.R
\docType{data}
\name{dynbenchmark_data}
\alias{dynbenchmark_data}
\title{The results data frame from dynbenchmark.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 159 columns.
}
\usage{
dynbenchmark_data
}
\description{
To do: add citation
}
\keyword{datasets}
