% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyheatmap-package.R
\docType{package}
\name{funkyheatmap-package}
\alias{funkyheatmap}
\alias{funkyheatmap-package}
\title{funkyheatmap: Generating funky heatmaps for benchmarks}
\description{
A visualization library for generating benchmarking heatmaps.
}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com}

Authors:
\itemize{
  \item Wouter Saelens \email{wouter.saelens@gmail.com}
}

}
\keyword{internal}
