% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_minmax.R
\name{scale_minmax}
\alias{scale_minmax}
\title{Scale a vector to the range [0, 1]}
\usage{
scale_minmax(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector scaled to the range [0, 1]
}
\description{
Scale a vector to the range [0, 1]
}
\examples{
scale_minmax(c(1, 2, 3))
}
