% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scib_data.R
\docType{data}
\name{scib_summary}
\alias{scib_summary}
\title{Summary results from the scIB project}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 27 columns.
}
\usage{
scib_summary
}
\description{
This dataset was generated by running the \code{data-raw/scib_summary.R} script.
This script downloads data for the RNA tasks from the
\href{https://github.com/theislab/scib-reproducibility}{scIB reproducibility repository}
and summarises it similarly (but not exactly like) what was done in the
original paper to give a final ranking of the top performing methods.
It is used in the scIB vignette (\code{vignette("scIB")}) to reproduce
the overall summary figure in Luecken et al. 2021.
}
\references{
{
\insertRef{scib_Luecken2021}{funkyheatmap}
}
}
\keyword{datasets}
