% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_data.R
\name{verify_data}
\alias{verify_data}
\title{Verify the integrity of the data object}
\usage{
verify_data(data)
}
\arguments{
\item{data}{A data frame with items by row and features in the columns.
Must contain one column named \code{"id"}.}
}
\value{
A verified data object
}
\description{
Verify the integrity of the data object
}
\examples{
library(tibble)
data <- tribble(
  ~id, ~name, ~x, ~y,
  "foo", "Foo", 0.5, 0.7,
  "bar", "Bar", 1.0, 0.1
)
verify_data(data)
}
