% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_legends.R
\name{create_rect_legend}
\alias{create_rect_legend}
\title{Create a rect legend}
\usage{
create_rect_legend(title, palette, position_args = position_arguments())
}
\description{
Create a rect legend
}
\examples{
title <- "Greys"
palette <- funkyheatmap:::default_palettes$numerical$Greys
create_rect_legend(title, palette)
}
