% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynbenchmark_data.R
\docType{data}
\name{dynbenchmark_data}
\alias{dynbenchmark_data}
\title{The results data frame from dynbenchmark.}
\format{
An object of class \code{list} of length 6.
}
\usage{
dynbenchmark_data
}
\description{
This data was generated by running the \code{data-raw/dynbenchmark_data.R} script.
It is used in the vignette named "\code{vignette("dynbenchmark")}" to regenerate
the results figures in Saelens et al. 2019.
}
\references{
\insertRef{comparisonsinglecell_saelens2019}{funkyheatmap}
}
\keyword{datasets}
