% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyheatmap-package.R
\docType{package}
\name{funkyheatmap-package}
\alias{funkyheatmap}
\alias{funkyheatmap-package}
\title{funkyheatmap: Generating Funky Heatmaps for Data Frames}
\description{
Allows generating beautiful visualisations for benchmark data frames. Funky heatmaps can be fine-tuned by providing annotations of the columns and rows, which allows assigning multiple palettes or geometries or grouping rows and columns together in categories.

Allows generating beautiful visualisations for benchmark data
frames. Funky heatmaps can be fine-tuned by providing annotations of the
columns and rows, which allows assigning multiple palettes or geometries
or grouping rows and columns together in categories.
}
\details{
This package heavily depends on the functions provided by
\href{https://ggplot2.tidyverse.org/}{\code{ggplot2}} and
\href{https://ggforce.data-imaginist.com}{\code{ggforce}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rcannood.github.io/funkyheatmap}
  \item \url{https://github.com/rcannood/funkyheatmap}
  \item Report bugs at \url{https://github.com/rcannood/funkyheatmap/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood)

Authors:
\itemize{
  \item Wouter Saelens \email{wouter.saelens@gmail.com} (\href{https://orcid.org/0000-0002-7114-6248}{ORCID}) (zouter)
}

}
\keyword{internal}
