% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyheatmap-package.R
\docType{package}
\name{funkyheatmap-package}
\alias{funkyheatmap}
\alias{funkyheatmap-package}
\title{Generating funky heatmaps for benchmarks}
\description{
Allows generating heatmap-likes visualisations for benchmark data
frames. Funky heatmaps can be fine-tuned by providing annotations of the
columns and rows, which allows assigning multiple palettes or geometries
or grouping rows and columns together in categories.
}
\details{
To learn more about funkyheatmap, please visit the
\href{https://funkyheatmap.github.io/funkyheatmap}{project page}.

This package heavily depends on the functions provided by
\href{https://ggplot2.tidyverse.org/}{\code{ggplot2}} and
\href{https://ggforce.data-imaginist.com}{\code{ggforce}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://funkyheatmap.github.io/funkyheatmap/}
  \item \url{https://github.com/funkyheatmap/funkyheatmap}
  \item Report bugs at \url{https://github.com/funkyheatmap/funkyheatmap/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood)

Authors:
\itemize{
  \item Wouter Saelens \email{wouter.saelens@gmail.com} (\href{https://orcid.org/0000-0002-7114-6248}{ORCID}) (zouter)
}

Other contributors:
\itemize{
  \item Louise Deconinck \email{louise.deconinck@gmail.com} (\href{https://orcid.org/0000-0001-8100-6823}{ORCID}) (LouiseDck) [contributor]
  \item Artuur Couckuyt (\href{https://orcid.org/0000-0001-7858-6521}{ORCID}) [contributor]
  \item Nick Markov (\href{https://orcid.org/0000-0002-3659-4387}{ORCID}) (mxposed) [contributor]
  \item Luke Zappia \email{luke@lazappi.id.au} (\href{https://orcid.org/0000-0001-7744-8565}{ORCID}) (lazappi) [contributor]
}

}
\keyword{internal}
